+++ src/nmem/analysis/array_fidelity_plot.py
